pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ssschem"
version = "0.1.0"
description = "State-space schematic generator producing signal-flow graphs and integrator diagrams."
readme = "README.md"
requires-python = ">=3.11"
license = {file = "LICENSE"}
authors = [{name = "State-Space Tools"}]
dependencies = [
    "sympy>=1.12",
    "numpy>=1.26",
    "networkx>=3.2",
    "pydot>=1.4",
    "pyyaml>=6.0",
    "jsonschema>=4.19",
    "typer>=0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.3",
    "mypy>=1.8",
    "types-PyYAML>=6.0",
]

[project.scripts]
ssschem = "ssschem.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100
select = ["E", "F", "I"]

[tool.pytest.ini_options]
minversion = "7.4"
addopts = "-ra"
testpaths = ["tests"]
